requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.10
django-extensions==3.2.3
djangorestframework==3.15.2
sqlparse==0.4.4
typing-extensions==4.9.0
python-dotenv==0.19.1
requests==2.26.0

